refactor: LottoNumberMakers의 로또 숫자 생성 기능을 LottoNumbers 내부 메서드로 변경 후 제거
